Fix Retrofit2 templates for Play framework
In the API templates for Retrofit2 with Play framework, api.mustache,
retrofit2.Response is imported but it is conflicted with the definition
of the response defined inline in the responses section and it does not
refer to the external definition by "ref" pointer. This causes the
compilation errors since the generated code are imported before
retrofit2.Response and they do not get along with the definition of
Retrofit2's Response class.

This patch eliminates the import statements of retrofit2.Response and
replace Response uses with fully qualified class name.

Signed-off-by: Taku Fukushima <f.tac.mac@gmail.com>